move additional func to helpers module
